e-bOOK GENERATOR/requirements.txt
Flask==2.3.3
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
fpdf2==2.7.8
beautifulsoup4==4.12.2
reportlab==4.4.4
